Move createFontManager() code to sun.font.PlatformFontInfo class
