+++ src/java.desktop/macosx/classes/sun/awt/PlatformGraphicsInfo.java
@@ -29,8 +29,6 @@ import java.awt.GraphicsEnvironment;
 import java.awt.Toolkit;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
-import sun.font.FontManager;
-import sun.font.CFontManager;
 
 @SuppressWarnings("removal")
 public class PlatformGraphicsInfo {
@@ -50,9 +48,6 @@ public class PlatformGraphicsInfo {
         return new sun.lwawt.macosx.LWCToolkit();
     }
 
-    public static FontManager createFontManager() {
-        return new CFontManager();
-    }
     /**
      * Returns true if the WindowServer is available, false otherwise.
      *
