+++ src/java.desktop/macosx/classes/sun/font/PlatformFontInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,36 +23,15 @@
  * questions.
  */
 
-package sun.awt;
+package sun.font;
 
-import java.awt.GraphicsEnvironment;
-import java.awt.Toolkit;
+class PlatformFontInfo {
 
-public class PlatformGraphicsInfo {
-
-    public static GraphicsEnvironment createGE() {
-        return new Win32GraphicsEnvironment();
-    }
-
-    public static Toolkit createToolkit() {
-        return new sun.awt.windows.WToolkit();
-    }
-
-    public static boolean getDefaultHeadlessProperty() {
-        // On Windows, we assume we can always create headful apps.
-        // Here is where we can add code that would actually check.
-        return false;
-    }
-
-    /*
-     * Called from java.awt.GraphicsEnvironment when
-     * getDefaultHeadlessProperty() has returned true, and
-     * the application has called an API that requires headful.
+    /**
+     * The method is only to be called via the
+     * {@code FontManagerFactory.getInstance()} factory method.
      */
-    public static String getDefaultHeadlessMessage() {
-        return
-            "\nThe application does not have desktop access,\n" +
-            "but this program performed an operation which requires it.";
+    static FontManager createFontManager() {
+        return new CFontManager();
     }
-
 }
