+++ src/java.desktop/unix/classes/sun/awt/PlatformGraphicsInfo.java
@@ -29,8 +29,6 @@ import java.awt.GraphicsEnvironment;
 import java.awt.Toolkit;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
-import sun.font.FontManager;
-import sun.awt.X11FontManager;
 
 public class PlatformGraphicsInfo {
 
@@ -40,10 +38,6 @@ public class PlatformGraphicsInfo {
 
     public static Toolkit createToolkit() {
         return new sun.awt.X11.XToolkit();
-    }
-
-    public static FontManager createFontManager() {
-        return new X11FontManager();
     }
 
     /**
